+++ src/lib.rs
@@ -68,7 +68,10 @@
 //! [Apache License 2.0](https://github.com/Qiskit/qiskit/blob/main/LICENSE.txt)
 
 #![warn(missing_docs)]
+/// The circuit module
+pub mod circuit;
 /// The main qiskit-rs module
-pub mod qiskit;
+mod qiskit;
 
-pub use qiskit::{ClassicalRegister, QiskitError, QuantumCircuit, QuantumRegister};
+pub use circuit::{ClassicalRegister, QuantumCircuit, QuantumRegister};
+pub use qiskit::*;
